Fixing some build issues in makefile by pulling the object files out of the line for building server into a variable so I can refer to them in the command to build server.
Updated the README with the simplest viable product. Working
toward that now.

Added the ability to scan the current directory looking for images
when the server boots up to server.c. Also fixed some issues in
server.c and added a param to print_blob that lets me limit the
total number of lines to prevent massive files from going beyond
the terminal's scrollback buffer.