Adding doctype to asl.html.
Trying to figure out why the 94.jpeg image isn't loading on the
page and I found out that the server is just sending a handful of
bytes. So I'm reworking the send_data function to not use string
placement with snprintf since the data might be a file and switching
to memcpy. This fails right now though as the buffer I have to
store the data is only a megabyte. I might need to send the
buffer in pieces: first the header, then the body.