Fixed the param parsing issue in POST.
It was the end of the header/parameter separation delimiter that
was causing the issue. "\r\n\r\n" is the end of the header, but
the POST parameters follow that in the message sent to the server.
So I updated the parse_request_buffer function to trim the
"rest_of_header" variable to end at \r\n\r\n.

Next I should look and make sure it parses the POST params
correctly.