Restructured the code.
Moved a bunch of functions that dealth with the http request out
of crvr.c and into http.c. That way I can reuse functions that
parse the header params to parse the post params.

Has a bug though and isn't finding any post params. But I think
it might be looking for ": " as the delimiter between params,
which is the case for header params, but post params use "=" as
the delimiter. Need to check on that.