Fixed the server waiting for data.
The server was always expecting the content-length in addition
to the current buffer the client sent. But if the client sent
everything, the server would go back to read the content length
again, which would hang the server and the client would never
send the information the server wanted.

I updated (repaired? since this worked before...) the code so
the server computes whether the buffer plus content length was
already received. If it was it skips the read for more data,
since there isn't any.

This seems to work so far. Will continue testing.