Working on loading the text_quizzer plugin.
Added the plugins directory and put the text_quizzer in there.

Update the makefile to build specific plugins by name. It will
look for a C file in the plugins dir to build for it. This creates
a dynamic library.

Updated plugins.c to append the correct extension to the dynamic
library being loaded from the plugins structure. This forced me to
add a config.h file. Although I can retrofit this in the future to
select the right one based on the config.mk.

Finally populated text_quizzer with a load_plugin function and I
see that crvr is calling it. I just need to add the other functions
to the text_quizzer for it to load successfully.