Adding html to text_quizzer.
Finished writing most of text_quizzer. But I need to get the html
pages converted to C and added to the quizzer so I don't need to
carry the html around the whole time. So wrote out the html pages
and started on the html2c tool, which will convert html pages to
strings in a header file. Need to keep working on that though.