Improving the quizzer.
Added <pre> tags around the questions and answer so they are
formatted as the author intended. This works for now but it might
be nice to think about adding html bullets and stuff possibly.
But it would require guessing the author's intent.

Also added support to use asterisks and hypens for indentation.